# Rediscuss.Microservices/services/Rediscuss.ForumService/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/follows": {
      "post": {
        "operationId": "FollowsController_follow",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFollowDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetFollowDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Follow a discuit",
        "tags": [
          "follows"
        ]
      }
    },
    "/follows/unfollow": {
      "post": {
        "operationId": "FollowsController_unfollow",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemoveFollowDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetFollowDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Unfollow a discuit",
        "tags": [
          "follows"
        ]
      }
    },
    "/follows/user/{userId}": {
      "get": {
        "operationId": "FollowsController_getUserFollows",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetFollowDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get follows for a user",
        "tags": [
          "follows"
        ]
      }
    },
...
  },
  "info": {
    "title": "Rediscuss Forum Api",
    "description": "API for Rediscuss Forum Application",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "GetDiscuitDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdBy": {
            "type": "number"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedBy": {
            "type": "number"
          },
          "CreatedByUsername": {
            "type": "string"
          },
          "UpdatedByUsername": {
            "type": "string"
          },
          "followersCount": {
            "type": "number"
          },
          "isFollowedByCurrentUser": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "CreateFollowDto": {
        "type": "object",
        "properties": {
          "discuitId": {
            "type": "string",
            "description": "Discuit id to follow"
          },
          "userId": {
            "type": "number",
            "description": "User id who follows"
          }
        },
        "required": [
          "discuitId",
          "userId"
        ]
      },
      "GetFollowDto": {
        "type": "object",
        "properties": {
          "discuitId": {
            "type": "string"
          },
          "userId": {
            "type": "number"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "discuit": {
            "$ref": "#/components/schemas/GetDiscuitDto"
          }
        },
        "required": [
          "discuitId",
          "userId",
          "discuit"
        ]
      },
      "RemoveFollowDto": {
        "type": "object",
        "properties": {
          "discuitId": {
            "type": "string",
            "description": "Discuit id to unfollow"
          },
          "userId": {
            "type": "number",
            "description": "User id who unfollows"
          }
        },
        "required": [
          "discuitId",
          "userId"
        ]
      },
...
    }
  }
}